+++ doc/changelog.rst
@@ -6,6 +6,8 @@ Changelog
 - Support for asyncio in Python 3. For each trollius module there is now an
   equivalent asyncio module. The installed utilities use asyncio on Python
   3.4+.
+- Add :attr:`spead2.recv.Stream.stop_on_stop_item` to allow a stream to keep
+  receiving after a stop item is received.
 - Switch shutdown code to use atexit instead of a capsule destructor, to
   support PyPy.
 - Test PyPy support with Travis.